ImageModel implementation. Added new fields with post image dimensions

+++ src/main/kotlin/net/glazov/data/model/ImageModel.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ImageModel(
     val imageUrl: String,
-    val imageWidth: Float,
-    val imageHeight: Float,
+    val imageWidth: Int,
+    val imageHeight: Int,
 )
